refactor: reorganize HSM component structure and add media preview functionality
